add cmake support for linux/mac/windows
This is a complete reproduction of the autotools code in cmake, with
improvements where cmake has an advantage.

The options are documented in README.md

Dependencies on Windows with Visual Studio  are handled with vcpkg, and
the `VCPKG_TARGET_TRIPLET` variable can activate vcpkg support on other
platforms as well.

Include the FindPython2 cmake modules as they are not in the base cmake
package on Ubuntu.

Make some minor fixes to the autotools code as well, for things broken
by adding the cmake support and some other fixes.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>